Main mechanic ready
You can pull, drag and drop the trash, and it checks if it's the correct garbage can or not (for now you can only see the type of the trash in the editor). Spawn the trash with " E ". There's a couple sounds implemented as well, but for some reason my Unity isn't playing any audio at all